requirements.txt
arrow==0.12.1
decorator==4.2.1
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
louvain==0.6.1
numpy==1.14.1
parso==0.1.1
pbr==3.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-igraph==0.7.1.post6
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
stevedore==1.28.0
traitlets==4.3.2
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wcwidth==0.1.7
